Solve issue when tapping on a video
The following method `-(void)handelImageTap:(UIGestureRecognizer *)gestureRecognizer` in MHGalleryImageViewerViewController was crashing my app when tapping on a video presented inside the `MHGalleryController`. The problem was due to the selector `locationInView:` called on an instance of UIButton. 

The `handelImageTap` method should actually accept only instances of UIGestureRecognizer as its parameters, but in `changeToPlayable` the `self.moviewPlayerButtonBehinde` is being passed as a parameter, thus crashing the app. 

I've solved the issue by adding a gesture recognizer to `self.moviewPlayerButtonBehinde` and by adding a defensive check inside `handelImageTap`.