sharedManager renamed in sharedInstance to add compatibility to swift
if you use this library in a swift project, you can't access
MHGallerySharedManager.sharedManager().
The compiler complains because it can't see sharedManager.
I looked around and find a fix thanks to this article
http://justinmiller.io/posts/2015/01/28/swift-objc-initializers/
So if you rename sharedManager in sharedInstance you are able to access
the singleton instance like this:
MHGallerySharedManager.sharedInstance().webPointForThumb