Thumbnail and duration
The original implementation was did with API v2 which are deprecated
right now. I replaced the old implementation with the new that requires
an API key generated from Google Api Console, this key must be
specified at row 6 of MHGallery.m through the const string
MHYoutubeApiKey. Really, only the duration need to be called via v3 API
because the thumbnails are available without API call (a little trick).
I know that’s a quick and dirty solution in fact, the rest part of
library will have to be modified.